Refactor for MVVM
Refactoring the code to provide better separation between the View,
Model and ViewModel. Since RichTextBox can't do data binding directly,
adding code to extract the plain text and pass it along to the ViewModel
manually. Moving the parsing logic to the model, as well as creating a
representation that keeps track of what are hashtags and what are
usernames. Coding the ViewModel to represent this data as simply
highlighted or not highlighted when it presents it to the View.